+++ src/main/java/com/mrpowergamerbr/loritta/userdata/LorittaServerUserData.kt
@@ -4,7 +4,7 @@ package com.mrpowergamerbr.loritta.userdata
  * Usado para guardar informações especificas para um servidor, como XP, nicknames, etc.
  */
 class LorittaServerUserData {
-	var xp: Int = 0; // XP do usuário no servidor
+	var xp: Long = 0; // XP do usuário no servidor
 	var isMuted: Boolean = false
 	var temporaryMute: Boolean = false
 	var expiresIn: Long = 0L
